extract ui storage and auth helpers
